DRF/aiogram/requirements.txt
requests==2.28.1
python-decouple==3.6
aiogram==2.15
